pyproject.toml
[project]
name = "diff-anbn"
version = "0.1.0"
description = "Masked Diffusion Language Models for Toy Formal Languages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
    "x-transformers>=1.30",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diff_anbn"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
